WIP: Fix blitting framebuffers with out-of-range dimensions on NVIDIA GL
NVIDIA GL drivers have an issue where blitFramebuffer fails if the
computed dimensions are outside the valid 32-bit integer range. We now
have a workaround for handling them as a special case in the GL
backend.

The workaround can take two different paths. In case the source and
destination dimensions match, the blitted area can be simply clipped
so that it lies entirely in the positive integer range.

In case the source and destination dimensions don't match, the shader-
based blit is used instead of calling glBlitFramebuffer. The shader-
based blit is updated so that it can clip the destination area to the
destination framebuffer and transform the source area with the same
transform. This is done in case the dimensions are too large to be
used as viewport dimensions.

Transforming the source area needs to be done using double precision
floats for the intermediate values. This way they can accurately
represent all 32-bit integers and numerical errors can't lead to
incorrect blit boundaries. The included end2end tests for the
workaround would fail in case only single precision were to be used.

The shader-based GL blit is is also fixed to handle dimensions
greater than max 32-bit int correctly. To do this, the Rectangle type
is updated to store width and height as 64-bit integers instead of
32-bit.

BUG=chromium:830046
TEST=angle_end2end_tests

+++ src/libANGLE/renderer/d3d/d3d11/StateManager11.cpp
@@ -1271,8 +1271,10 @@ void StateManager11::syncScissorRectangle(const gl::Rectangle &scissor, bool ena
             int y            = modifiedScissorY + mViewportOffsets[i].y;
             rect.left        = std::max(0, x);
             rect.top         = std::max(0, y);
-            rect.right       = x + std::max(0, scissor.width);
-            rect.bottom      = y + std::max(0, scissor.height);
+            // Scissor width and height have been specified as 32-bit ints so it's safe to downcast
+            // them.
+            rect.right  = x + std::max(0, static_cast<int>(scissor.width));
+            rect.bottom = y + std::max(0, static_cast<int>(scissor.height));
         }
         mRenderer->getDeviceContext()->RSSetScissorRects(numRectangles, rectangles.data());
     }
@@ -1350,10 +1352,12 @@ void StateManager11::syncViewport(const gl::Context *context)
         if (!framebuffer->getFirstNonNullAttachment() &&
             (framebuffer->getDefaultWidth() || framebuffer->getDefaultHeight()))
         {
-            dxViewport.Width =
-                static_cast<GLfloat>(std::min(viewport.width, framebuffer->getDefaultWidth()));
-            dxViewport.Height =
-                static_cast<GLfloat>(std::min(viewport.height, framebuffer->getDefaultHeight()));
+            // Viewport width and height have been specified as 32-bit ints so it's safe to downcast
+            // them.
+            dxViewport.Width = static_cast<GLfloat>(
+                std::min(static_cast<int>(viewport.width), framebuffer->getDefaultWidth()));
+            dxViewport.Height = static_cast<GLfloat>(
+                std::min(static_cast<int>(viewport.height), framebuffer->getDefaultHeight()));
         }
         else
         {
